widl: Allow importing wine sources from an out of tree wine build
Some of the imported source files are autogenerated, and if wine
was built outside of the source tree, the caller of wine-import.sh
needs to provide paths to both trees.

In such an import, the generated bison/flex files will have stray
changes due to differing source paths in preprocessor outputs though.

+++ mingw-w64-tools/widl/wine-import.sh
@@ -3,9 +3,10 @@
 set -e
 
 WINE_DIR=$1
+WINE_BUILD_DIR=$2
 
 if test -z $WINE_DIR; then
-	echo "Usage: wine-import.sh wine_dir"
+	echo "Usage: wine-import.sh wine_dir [wine_build_dir]"
 	exit 1
 fi
 
@@ -13,12 +14,20 @@ for f in client.c expr.c expr.h hash.c hash.h header.c header.h parser.h parser.
 	parser.yy.c proxy.c register.c server.c typegen.c typegen.h typelib.c typelib.h typelib_struct.h typetree.c typetree.h \
 	utils.c utils.h widl.c widl.h widltypes.h write_msft.c
 do
-	cp $WINE_DIR/tools/widl/$f src
+	if [ ! -f $WINE_DIR/tools/widl/$f ] && [ -f $WINE_BUILD_DIR/tools/widl/$f ]; then
+		cp $WINE_BUILD_DIR/tools/widl/$f src
+	else
+		cp $WINE_DIR/tools/widl/$f src
+	fi
 done
 
 for f in ppl.l ppl.yy.c ppy.tab.c ppy.tab.h ppy.y preproc.c wpp.c wpp_private.h
 do
-	cp $WINE_DIR/libs/wpp/$f src/wpp
+	if [ ! -f $WINE_DIR/libs/wpp/$f ] && [ -f $WINE_BUILD_DIR/libs/wpp/$f ]; then
+		cp $WINE_BUILD_DIR/libs/wpp/$f src/wpp
+	else
+		cp $WINE_DIR/libs/wpp/$f src/wpp
+	fi
 done
 
 for f in basetsd.h guiddef.h poppack.h pshpack1.h pshpack2.h pshpack4.h pshpack8.h winbase.h windef.h winerror.h winnls.h winnt.h winnt.rh
